ethereum-staking-pipeline/api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
snowflake-connector-python==3.5.0
python-dotenv==1.0.0
